Change the mutator methods of StreamGraph to default scope
This is minimal effort approach to make the StreamGraph immutable. It is an attempt to revert to a state when the StreamGraphGenerator was the only entry point to configure the StreamGraph.

Making the StreamGraphGenerator the only place where the StreamGraph is constructed makes it easier to reason about the properties of the StreamGraph.