[hotfix] Do not cache DataOutput in StatusWatermarkValve
StatusWatermarkValve was caching a DataOutput passed in a constructor. From an architecture point of view it is a bug, because it ignores the DataOutput passed in the DataInput#emitNext(DataOutput). It made and implicit assumption that these two are always equal, which might not be true.

In this PR I am passing the DataOutput as a parameter of StatusWatermarkValve methods.