[FLINK-20491] Turn BroadcastStateTransformation into "logical" Transformation
Note: Broadcast operations in BATCH mode don't yet work with this
change. This needs follow-up changes from later commits. We just lay the
groundwork here and keep the same functionality.

Before, we were creating operators in BroadcastConnectedStreams eagerly.
Now, the transformation holds the user function and we add a Translator
that creates the "physical" operators when translating the graph of
Transformations.

We do this so that we can translate differently based on whether we're
in BATCH or STREAMING mode.