[FLINK-22686][checkpointing] Incompatible subtask mappings while resuming from unaligned checkpoints
In the initial implementation of rescaling channel data we assumed that
the subtasks indices are the same for all input gates. This assumptions
does not hold when there are different partitioners. It can happen e.g.
in case of a broadcast with a keyed input.

The commit tracks the subtask indices separately for each partition and
input gate.

This closes #16019