[XXX][source] Add AlignedSourceReader for aligned SourceReaders.
Implementors of a source can choose to use an AlignedSourceReader, such that all subtasks are roughly aligned on the event time. The AlignedSourceReader is meant as a specialized interface that only sources where splits simultaneously progress on watermarks. (It's not applicable to standard file sources where splits have no strict watermark progression)