[hotfix] Make LocalFileSystem#listStatus resilient to concurrent deletes
LocalFileSystem#listStatus calls File#list to retrieve list of file
names inside of a directory. After that it calls
LocalFileSystem#getFileStatus to convert a previously retrieved file
name to a LocalFileStatus. However between these two calls a file from
the directory could've been removed which results in the getFileStatus
throwing a FileNotFoundException. Instead of calling getFileStatus we
can convert the file name into a LocalFileStatus in place after the
atomic File#list.