[FLINK-25745] Deleting claimed storage location
Savepoint is self contained and relocatable. At the same time once we
support incremental savepoints it can contain files that can be shared
with subsequent checkpoints (if restored in claim mode). Such shared
files would reside in the savepoint location. Because of that fact, we
can not simply dispose its location when the savepoint is being
subsumed.

To overcome this, we can tell CompletedCheckpointStore to try to dispose
such a location, whenever it subsumes any checkpoint, because such
operation can delete the last shared file inside of that location.